+++ src/animation/controllers/index.ts
@@ -5,3 +5,4 @@ export { withEase } from './withEase';
 export { withSequence } from './withSequence';
 export { withSpring } from './withSpring';
 export { withTiming } from './withTiming';
+export { withLoop } from './withLoop';