+++ src/animation/controllers/withSpring.ts
@@ -1,10 +1,14 @@
 import { AnimationConfig } from '../animationType';
 import { type UseValueConfig } from '../hooks';
 import { withConfig, type WithOnCallbacks } from './withConfig';
+import { UpdateValue } from '../core/FluidController';
 
 interface WithSpringConfig
   extends Pick<UseValueConfig, 'mass' | 'friction' | 'tension'>,
     WithOnCallbacks {}
 
-export const withSpring = (toValue: number, config?: WithSpringConfig) =>
+export const withSpring = (
+  toValue: number,
+  config?: WithSpringConfig
+): UpdateValue =>
   withConfig(toValue, { ...AnimationConfig.ELASTIC, ...config });